refactoring save query and redirect code
